Refactor: Relocate series data source and models
This commit relocates the `SeriesRemoteDataSource` interface and the `SeriesDto` and `GenreDto` data classes from the `datasource/remote/series` module to the `repository/series` module.

Additionally, the following changes were made:

- Updated `TmdbSeriesApiRemoteDataSource` to use the new location of `SeriesRemoteDataSource` and the model classes.
- Modified `TmdbSeriesApiRemoteDataSource` to accept `genreId` as `Int` instead of `String`.
- Updated `BaseRequest` to accept `baseURL` and `accessToken` in its constructor and removed unused methods.
- Updated `build.gradle.kts` files to reflect the new module dependencies and plugin usage.
- Renamed Koin module from `seriesDataModule` to `seriesRemoteDataModule`.
- Adjusted Koin module to retrieve `BASE_URL` and `ACCESS_TOKEN` using named qualifiers.
- Removed explicit `@SerialName` annotations from DTOs where the property name matches the JSON key.
- Reordered fields in `SeriesDto`.
