+++ app/src/main/java/com/giraffe/cineverseapp/CineVerseApp.kt
@@ -2,7 +2,7 @@ package com.giraffe.cineverseapp
 
 import android.app.Application
 import com.giraffe.cineverseapp.di.dataModule
-import com.giraffe.series.di.seriesDataModule
+import com.giraffe.series.di.seriesRemoteDataModule
 import com.giraffe.cineverseapp.di.localDataSourceModule
 import com.giraffe.cineverseapp.di.networkModule
 import com.giraffe.cineverseapp.di.repositoryModule
@@ -21,7 +21,7 @@ class CineVerseApp : Application() {
                 networkModule,
                 repositoryModule,
                 useCaseModule,
-                seriesDataModule
+                seriesRemoteDataModule
             )
         }
     }
