+++ datasource/remote/series/src/main/java/com/giraffe/series/TmdbSeriesApiRemoteDataSource.kt
@@ -2,9 +2,10 @@ package com.giraffe.series
 
 import com.giraffe.series.api.BaseRequest
 import com.giraffe.series.api.RequestBuilder
-import com.giraffe.series.model_dto.GenreDto
+import com.giraffe.series.datasource.remote.SeriesRemoteDataSource
+import com.giraffe.series.model.GenreDto
+import com.giraffe.series.model.SeriesDto
 import com.giraffe.series.model_dto.GenresResponse
-import com.giraffe.series.model_dto.SeriesDto
 import com.giraffe.series.model_dto.SeriesResponse
 import io.ktor.http.HttpMethod
 
@@ -37,7 +38,7 @@ internal class TmdbSeriesApiRemoteDataSource(
     }
 
     override suspend fun getSeriesByGenreId(
-        genreId: String,
+        genreId: Int,
         language: String,
         includeAdult: Boolean,
         page: Int,
