+++ datasource/remote/series/src/main/java/com/giraffe/series/di/SeriesDataModule.kt
@@ -1,20 +1,21 @@
 package com.giraffe.series.di
 
-import com.giraffe.series.BuildConfig
-import com.giraffe.series.SeriesRemoteDataSource
 import com.giraffe.series.TmdbSeriesApiRemoteDataSource
 import com.giraffe.series.api.BaseRequest
 import com.giraffe.series.api.RequestBuilder
+import com.giraffe.series.datasource.remote.SeriesRemoteDataSource
 import org.koin.core.module.dsl.singleOf
+import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
 
-val seriesDataModule = module {
-
-    single { BuildConfig.TMDB_API_KEY }
-
-    singleOf(::BaseRequest)
-
+val seriesRemoteDataModule = module {
+    single {
+        BaseRequest(
+            baseURL = get(named("BASE_URL")),
+            accessToken = get(named("ACCESS_TOKEN"))
+        )
+    }
     singleOf(::RequestBuilder)
 
     single<SeriesRemoteDataSource> {
