+++ datasource/remote/series/src/main/java/com/giraffe/series/model_dto/GenresResponse.kt
@@ -1,10 +1,9 @@
 package com.giraffe.series.model_dto
 
-import kotlinx.serialization.SerialName
+import com.giraffe.series.model.GenreDto
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class GenresResponse(
-    @SerialName("genres")
     val genres: List<GenreDto>
 )
