+++ repository/series/src/main/java/com/giraffe/series/model/SeriesDto.kt
@@ -1,31 +1,26 @@
-package com.giraffe.series.model_dto
+package com.giraffe.series.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class SeriesDto(
-    @SerialName("id")
     val id: Int,
-    @SerialName("name")
     val name: String,
+    val voteCount: Int,
+    val overview: String,
+    val popularity: Double,
     @SerialName("original_name")
     val originalName: String,
-    @SerialName("overview")
-    val overview: String,
     @SerialName("first_air_date")
     val firstAirDate: String,
     @SerialName("poster_path")
     val posterPath: String?,
     @SerialName("backdrop_path")
     val backdropPath: String?,
-    @SerialName("popularity")
-    val popularity: Double,
     @SerialName("vote_average")
     val voteAverage: Double,
     @SerialName("vote_count")
-    val voteCount: Int,
-    @SerialName("adult")
     val adult: Boolean = false,
     @SerialName("genre_ids")
     val genreIds: List<Int> = emptyList(),
